This commit does
1. Change the types of the integer max value constants.
2. Limit the max capacity of FE.Core scalable pool to 2^32 / 2 bytes. (MSVC constraints)
- Change the integer variable types [size_t, uint64, and int64] to int32 accordingly.

+++ SDK/Core/Include/FE/pool/private/pool_common.hxx
@@ -76,7 +76,7 @@ namespace internal::pool
     struct block_info
     {
         var::byte* _address;
-        var::int64 _size_in_bytes;
+        var::int32 _size_in_bytes;
     };
 
     template<PoolType PoolType, PoolPageCapacity PageCapacity, class Alignment>
